Minimus 0.0.8
Added a hook to allow Twisted web server.  Added a hook to SSL certificate with GEvent server. Added a separate `parse_json(env)` method.  Fixed a small bug with a blank web page.  Promoted `jsonify(dict)` that was part of the Minimus class to be called at the top level.